Install only qtcreator’s own docs, not also QBS docs.
The current install target tries to install QBS documentation, but
it is not built so there is an error. In Qt 5.12 that error is ignored.

However, in Qt 5.14 that error is no longer ignored because of this
upstream commit:

https://code.qt.io/cgit/qt/qtbase.git/commit/?id=20e9422e069a27b7

The `install_docs` target in the generated Makefile depends on three
subtargets: `install_inst_qch_docs`, `install_inst_html_docs` and
`install_qbs_docs`. This commit avoids calling the third by explicitly
calling the first two of them.

Also, it moves the docs install command into the ifneq block.
When building architecture-dependent packages, the documentation is not
built so trying to install it will also cause an error with Qt 5.14.